+++ src/id.rs
@@ -184,6 +184,7 @@ impl RawID {
 
 #[cfg(test)]
 mod tests {
+  #[allow(unused_imports)]
   use super::*;
 
   #[test]